PR Feedback: move ProviderWithCustomServer
